Scorekeeper!

+++ apps/backend/src/lib/schedule/cleaner.ts
@@ -1,24 +1,19 @@
 import { WithId } from 'mongodb';
 import * as db from '@lems/database';
-import { Event, RobotGameTable, JudgingRoom, Team } from '@lems/types';
+import { Event, JudgingRoom, Team } from '@lems/types';
 
 export const cleanEventData = async (event: WithId<Event>) => {
+  if (!(await db.deleteEventUsers(event._id)).acknowledged)
+    throw new Error('Could not delete users!');
+
   const oldEventState = await db.getEventStateFromEvent(event._id);
   if (oldEventState) {
     if (!(await db.deleteEventState(oldEventState)).acknowledged)
       throw new Error('Could not delete event state!');
   }
 
-  const oldTables = await db.getEventTables(event._id);
   const oldRooms = await db.getEventRooms(event._id);
   const oldTeams = await db.getEventTeams(event._id);
-
-  await Promise.all(
-    oldTables.map(async (table: WithId<RobotGameTable>) => {
-      if (!(await db.deleteTableMatches(table._id)).acknowledged)
-        throw new Error('Could not delete matches!');
-    })
-  );
 
   await Promise.all(
     oldRooms.map(async (room: WithId<JudgingRoom>) => {
@@ -47,6 +42,8 @@ export const cleanEventData = async (event: WithId<Event>) => {
     throw new Error('Could not delete tables!');
   if (!(await db.deleteEventRooms(event._id)).acknowledged)
     throw new Error('Could not delete rooms!');
+  if (!(await db.deleteEventMatches(event._id)).acknowledged)
+    throw new Error('Could not delete matches!');
   if (!(await db.deleteEventUsers(event._id)).acknowledged)
     throw new Error('Could not delete users!');
 };
